removed model LastAttempt; improved update_compas functionality
