+++ src/ets/management/commands/sync_weekly_compas.py
@@ -1,14 +1,10 @@
 ### -*- coding: utf-8 -*- ####################################################
 
-from optparse import make_option
-from datetime import datetime, timedelta
-
 from django.core.management.base import BaseCommand
-from django.db.utils import DatabaseError
 from django.conf import settings
 
-from ets.utils import update_compas_info
-from ets.models import Compas, ImportLogger
+from ets import utils
+from ets.models import Compas
 
 LOG_DIRECTORY = settings.LOG_DIRECTORY
 MINIMUM_AGE = 5
@@ -22,17 +18,8 @@ class Command(BaseCommand):
         --compas -- COMPAS station identifier (i.e. ISBX002 for example)
         
     """
-    option_list = BaseCommand.option_list + ( 
-        make_option('--compas', dest='compas', default='',
-            help='Tells the system to synchronize only this one compas station'),
-    )
-
     help = 'Import data from COMPAS stations'
     
-    def synchronize(self, compas):
-        """Exact method to proceed synchronization"""
-        update_compas_info(compas)
-
-    def handle(self, compas='', *args, **options):
+    def handle(self, *args, **options):
         compas = Compas.objects.get(pk='PSHX001')
-        self.synchronize(compas=compas.pk)
+        utils.update_compas(compas.pk, utils.import_partners, utils.import_places, utils.import_reasons)
